python-service/requirements.txt
openai-whisper
watchdog
torch
torchaudio
numpy
